Implemented card deck

+++ Poker/player.cpp
@@ -2,5 +2,5 @@
 
 void Player::Draw(void)
 {
-	_pge->FillCircle(_posX, _posY, 40, olc::BLACK);
+	_pge->FillCircle(static_cast<int32_t>(_posX), static_cast<int32_t>(_posY), 40, olc::BLACK);
 }
